+++ test/common/config.js
@@ -56,6 +56,10 @@ export default function () {
         ...(!!process.env.CODE_COVERAGE
           ? [`--plugin-path=${path.join(__dirname, 'fixtures', 'plugins', 'coverage')}`]
           : []),
+        '--logging.appenders.deprecation.type=console',
+        '--logging.appenders.deprecation.layout.type=json',
+        '--logging.loggers[0].name=elasticsearch.deprecation',
+        '--logging.loggers[0].appenders[0]=deprecation',
       ],
     },
     services,
